requirements.txt
sanic==24.6.0
SQLAlchemy==2.0.36
greenlet==3.1.1
asyncpg==0.30.0
alembic==1.14.0
PyJWT==2.10.1
